2HW/requirements.txt
apipkg==1.5
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
execnet==1.7.1
get==2019.4.13
idna==2.9
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
post==2019.4.13
public==2019.4.13
py==1.8.1
pyparsing==2.4.6
pytest==5.4.1
pytest-forked==1.1.3
pytest-xdist==1.31.0
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
webdriver-manager==2.3.0
